+++ docs/misc/changelog.rst
@@ -31,6 +31,7 @@ Others:
 - Fixed ``stable_baselines3/common/utils.py`` type hints
 - Fixed ``stable_baselines3/common/vec_envs/vec_transpose.py`` type hints
 - Fixed ``stable_baselines3/common/vec_env/vec_video_recorder.py`` type hints
+- Fixed ``stable_baselines3/common/save_util.py`` type hints
 
 Documentation:
 ^^^^^^^^^^^^^^